add new featuer

+++ user_service/src/main/java/com/example/demo/controller/UserController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/users")
 public class UserController {
     @GetMapping("/{id}")
-    public User getUser(@PathVariable Long id) {
+    public User getUser(@PathVariable Integer id) {
         return new User(id, "John Doe","john.dev@gmail.com","015408894");
     }
 }
