+++ user_service/src/main/java/com/example/demo/domain/User.java
@@ -16,7 +16,7 @@ import lombok.Setter;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "name")
     private String name;
@@ -27,25 +27,4 @@ public class User {
     @Column(name = "phone_number")
     private String phoneNumber;
 
-//    public User() {}
-//    public User(Long id, String name) {
-//        this.id = id;
-//        this.name = name;
-//    }
-//
-//    public Long getId() {
-//        return id;
-//    }
-//
-//    public void setId(Long id) {
-//        this.id = id;
-//    }
-//
-//    public String getName() {
-//        return name;
-//    }
-//
-//    public void setName(String name) {
-//        this.name = name;
-//    }
 }
